+++ src/fxr.ts
@@ -711,9 +711,11 @@ export enum ActionType {
    * - **Slot**: {@link ActionSlots.AppearanceAction Appearance}
    * - **Class**: {@link RichModel}
    * 
-   * Particle with a 3D model. Similar to {@link ActionType.Model Model}, but with some different options and seemingly no way to change the blend mode.
+   * Particle with a 3D model. Similar to {@link ActionType.Model Model}, but with some different options.
    * 
    * Some models only work properly with this action and not with the Model action for some unknown reason. A good example of this is the Carian greatsword model in Elden Ring (88300), which gets horribly stretched and distorted when used with the other action, but it works fine with this one.
+   * 
+   * The name is from Elden Ring's RTTI.
    */
   RichModel = 10015,
   /**
@@ -2807,19 +2809,19 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       interval: { default: 1, time: 2 },
-      perInterval: { default: 1 },
-      totalIntervals: { default: -1 },
+      perEmission: { default: 1 },
+      totalEmissions: { default: -1 },
       maxConcurrent: { default: -1, field: 1 },
       unk_ds3_f1_1: { default: 1, field: 1 },
     },
     games: {
       [Game.DarkSouls3]: {
         fields1: ['maxConcurrent','unk_ds3_f1_1'],
-        properties1: ['interval','perInterval','totalIntervals']
+        properties1: ['interval','perEmission','totalEmissions']
       },
       [Game.Sekiro]: {
         fields1: ['unk_ds3_f1_1'],
-        properties1: ['interval','perInterval','totalIntervals','maxConcurrent']
+        properties1: ['interval','perEmission','totalEmissions','maxConcurrent']
       },
       [Game.EldenRing]: Game.Sekiro,
       [Game.ArmoredCore6]: Game.Sekiro
@@ -2831,7 +2833,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       threshold: { default: 0.1, scale: 1 },
-      unk_ds3_p1_2: { default: -1 },
+      totalEmissions: { default: -1 },
       maxConcurrent: { default: -1, field: 1 },
       unk_ds3_p1_1: { default: 1 },
       unk_ds3_f1_1: { default: 1, field: 1 },
@@ -2840,11 +2842,11 @@ const ActionData: Record<string, ActionDataEntry> = {
     games: {
       [Game.DarkSouls3]: {
         fields1: ['maxConcurrent','unk_ds3_f1_1'],
-        properties1: ['threshold','unk_ds3_p1_1','unk_ds3_p1_2']
+        properties1: ['threshold','unk_ds3_p1_1','totalEmissions']
       },
       [Game.Sekiro]: {
         fields1: ['unk_ds3_f1_1','unk_sdt_f1_1'],
-        properties1: ['threshold','unk_ds3_p1_2','maxConcurrent']
+        properties1: ['threshold','totalEmissions','maxConcurrent']
       },
       [Game.EldenRing]: Game.Sekiro,
       [Game.ArmoredCore6]: Game.Sekiro
@@ -4012,7 +4014,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_sdt_f2_25: { default: 0, field: 2 },
       phaseFunction: { default: true, field: 0 },
       asymmetryParam: { default: 0.75, field: 2 },
-      falloffExponent: { default: 1, field: 2 },
+      adaptationExponent: { default: 0, field: 2 },
       unk_er_f2_29: { default: 1, field: 1 },
       unk_er_f2_30: { default: 1, field: 2 },
       unk_er_f2_31: { default: 1, field: 1 },
@@ -4028,13 +4030,13 @@ const ActionData: Record<string, ActionDataEntry> = {
       },
       [Game.Sekiro]: {
         fields1: Game.DarkSouls3,
-        fields2: ['unk_ds3_f2_0','jitterAndFlicker','jitterAcceleration','unk_ds3_f2_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','fadeOutTime','shadowDarkness','unk_ds3_f2_15','unk_ds3_f2_16','unk_ds3_f2_17','unk_ds3_f2_18','unk_ds3_f2_19','unk_ds3_f2_20','unk_ds3_f2_21','unk_ds3_f2_22','unk_ds3_f2_23','maxViewDistance','volumeDensity','unk_sdt_f2_25','phaseFunction','asymmetryParam','falloffExponent'],
+        fields2: ['unk_ds3_f2_0','jitterAndFlicker','jitterAcceleration','unk_ds3_f2_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','fadeOutTime','shadowDarkness','unk_ds3_f2_15','unk_ds3_f2_16','unk_ds3_f2_17','unk_ds3_f2_18','unk_ds3_f2_19','unk_ds3_f2_20','unk_ds3_f2_21','unk_ds3_f2_22','unk_ds3_f2_23','maxViewDistance','volumeDensity','unk_sdt_f2_25','phaseFunction','asymmetryParam','adaptationExponent'],
         properties1: Game.DarkSouls3,
         properties2: ['unk_ds3_p2_0','unk_ds3_p2_1','unk_sdt_p2_2','diffuseMultiplier','specularMultiplier']
       },
       [Game.EldenRing]: {
         fields1: Game.DarkSouls3,
-        fields2: ['unk_ds3_f2_0','jitterAndFlicker','jitterAcceleration','unk_ds3_f2_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','fadeOutTime','shadowDarkness','unk_ds3_f2_15','unk_ds3_f2_16','unk_ds3_f2_17','unk_ds3_f2_18','unk_ds3_f2_19','unk_ds3_f2_20','unk_ds3_f2_21','unk_ds3_f2_22','unk_ds3_f2_23','maxViewDistance','volumeDensity','unk_sdt_f2_25','phaseFunction','asymmetryParam','falloffExponent','unk_er_f2_29','unk_er_f2_30','unk_er_f2_31','unk_er_f2_32'],
+        fields2: ['unk_ds3_f2_0','jitterAndFlicker','jitterAcceleration','unk_ds3_f2_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','fadeOutTime','shadowDarkness','unk_ds3_f2_15','unk_ds3_f2_16','unk_ds3_f2_17','unk_ds3_f2_18','unk_ds3_f2_19','unk_ds3_f2_20','unk_ds3_f2_21','unk_ds3_f2_22','unk_ds3_f2_23','maxViewDistance','volumeDensity','unk_sdt_f2_25','phaseFunction','asymmetryParam','adaptationExponent','unk_er_f2_29','unk_er_f2_30','unk_er_f2_31','unk_er_f2_32'],
         properties1: Game.DarkSouls3,
         properties2: Game.Sekiro
       },
@@ -5297,7 +5299,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_er_f1_20: { default: 0, field: 1 },
       unk_er_f1_21: { default: 0, field: 1 },
       unk_er_f1_22: { default: 0, field: 1 },
-      unk_er_f1_23: { default: 0, field: 1 },
+      unkBlendMode: { default: 0, field: 1 },
       unk_ac6_f1_24: { default: 0, field: 2 },
       unk_ac6_f1_25: { default: -1, field: 2 },
       unk_ac6_f1_26: { default: -1, field: 2 },
@@ -5349,13 +5351,13 @@ const ActionData: Record<string, ActionDataEntry> = {
     },
     games: {
       [Game.EldenRing]: {
-        fields1: ['orientation','scaleVariationX','scaleVariationY','scaleVariationZ','uniformScale','unk_er_f1_5','unk_er_f1_6','dither','unk_er_f1_8','unk_er_f1_9','anibnd','animation','loopAnimation','animationSpeed','unk_er_f1_14','unk_er_f1_15','unk_er_f1_16','unk_er_f1_17','unk_er_f1_18','unk_er_f1_19','unk_er_f1_20','unk_er_f1_21','unk_er_f1_22','unk_er_f1_23','unk_er_f1_24','unk_er_f1_25'],
+        fields1: ['orientation','scaleVariationX','scaleVariationY','scaleVariationZ','uniformScale','unk_er_f1_5','unk_er_f1_6','dither','unk_er_f1_8','unk_er_f1_9','anibnd','animation','loopAnimation','animationSpeed','unk_er_f1_14','unk_er_f1_15','unk_er_f1_16','unk_er_f1_17','unk_er_f1_18','unk_er_f1_19','unk_er_f1_20','unk_er_f1_21','unk_er_f1_22','unkBlendMode','unk_er_f1_24','unk_er_f1_25'],
         fields2: ['unk_er_f2_0','unk_er_f2_1','unk_er_f2_2','unk_er_f2_3','bloomColor','unk_er_f2_8','unk_er_f2_9','unk_er_f2_10','unk_er_f2_11','unk_er_f2_12','unk_er_f2_13','minFadeDistance','minDistance','maxFadeDistance','maxDistance','minDistanceThreshold','maxDistanceThreshold','unk_er_f2_20','unk_er_f2_21','unk_er_f2_22','unk_er_f2_23','unk_er_f2_24','unkDepthBlend1','unkDepthBlend2','unk_er_f2_27','unk_er_f2_28','unk_er_f2_29','unk_er_f2_30','unk_er_f2_31','unk_er_f2_32','unk_er_f2_33','unk_er_f2_34','unk_er_f2_35','unk_er_f2_36','unk_er_f2_37'],
         properties1: ['model','sizeX','sizeY','sizeZ','rotationX','rotationY','rotationZ','angularSpeedX','angularSpeedMultiplierX','angularSpeedY','angularSpeedMultiplierY','angularSpeedZ','angularSpeedMultiplierZ','color1','color2','color3','unk_er_p1_16','unk_er_p1_17','rgbMultiplier2','unk_er_p1_19','unk_er_p1_20','offsetU','offsetV','speedU','speedMultiplierU','speedV','speedMultiplierV'],
         properties2: ['rgbMultiplier','alphaMultiplier','unk_er_p2_2','unk_er_p2_3','unk_er_p2_4','unk_er_p2_5','unk_er_p2_6']
       },
       [Game.ArmoredCore6]: {
-        fields1: ['orientation','scaleVariationX','scaleVariationY','scaleVariationZ','uniformScale','unk_er_f1_5','unk_er_f1_6','dither','unk_er_f1_8','unk_er_f1_9','anibnd','animation','loopAnimation','animationSpeed','unk_er_f1_14','unk_er_f1_15','unk_er_f1_16','unk_er_f1_17','unk_er_f1_18','unk_er_f1_19','unk_er_f1_20','unk_er_f1_21','unk_er_f1_22','unk_er_f1_23','unk_ac6_f1_24','unk_ac6_f1_25','unk_ac6_f1_26','unk_ac6_f1_27','unk_ac6_f1_28','unk_ac6_f1_29','unk_ac6_f1_30','unk_ac6_f1_31','unk_ac6_f1_32','unk_ac6_f1_33','unk_ac6_f1_34'],
+        fields1: ['orientation','scaleVariationX','scaleVariationY','scaleVariationZ','uniformScale','unk_er_f1_5','unk_er_f1_6','dither','unk_er_f1_8','unk_er_f1_9','anibnd','animation','loopAnimation','animationSpeed','unk_er_f1_14','unk_er_f1_15','unk_er_f1_16','unk_er_f1_17','unk_er_f1_18','unk_er_f1_19','unk_er_f1_20','unk_er_f1_21','unk_er_f1_22','unkBlendMode','unk_ac6_f1_24','unk_ac6_f1_25','unk_ac6_f1_26','unk_ac6_f1_27','unk_ac6_f1_28','unk_ac6_f1_29','unk_ac6_f1_30','unk_ac6_f1_31','unk_ac6_f1_32','unk_ac6_f1_33','unk_ac6_f1_34'],
         fields2: Game.EldenRing,
         properties1: ['model','sizeX','sizeY','sizeZ','rotationX','rotationY','rotationZ','angularSpeedX','angularSpeedMultiplierX','angularSpeedY','angularSpeedMultiplierY','angularSpeedZ','angularSpeedMultiplierZ','color1','color2','color3','unk_er_p1_16','unk_er_p1_17','rgbMultiplier2','unk_er_p1_19','unk_er_p1_20','offsetUV','speedUV','speedMultiplierUV'],
         properties2: Game.EldenRing
@@ -5831,7 +5833,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_sdt_f1_20: { default: 0, field: 2 },
       phaseFunction: { default: true, field: 0 },
       asymmetryParam: { default: 0.75, field: 2 },
-      falloffExponent: { default: 1, field: 2 },
+      adaptationExponent: { default: 0, field: 2 },
       unk_er_f1_24: { default: 1, field: 1 },
       unk_er_f1_25: { default: 1, field: 2 },
       unk_er_f1_26: { default: 1, field: 1 },
@@ -5847,11 +5849,11 @@ const ActionData: Record<string, ActionDataEntry> = {
         properties1: ['diffuseColor','specularColor','near','far','radiusX','radiusY','unk_ds3_p1_6','unk_ds3_p1_7']
       },
       [Game.Sekiro]: {
-        fields1: ['unk_sdt_f1_0','jitterAndFlicker','jitterAcceleration','unk_sdt_f1_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','shadowDarkness','unk_ds3_f1_3','unk_ds3_f1_4','fadeOutTime','unk_sdt_f1_16','unk_sdt_f1_17','unk_sdt_f1_18','volumeDensity','unk_sdt_f1_20','phaseFunction','asymmetryParam','falloffExponent'],
+        fields1: ['unk_sdt_f1_0','jitterAndFlicker','jitterAcceleration','unk_sdt_f1_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','shadowDarkness','unk_ds3_f1_3','unk_ds3_f1_4','fadeOutTime','unk_sdt_f1_16','unk_sdt_f1_17','unk_sdt_f1_18','volumeDensity','unk_sdt_f1_20','phaseFunction','asymmetryParam','adaptationExponent'],
         properties1: ['diffuseColor','specularColor','diffuseMultiplier','specularMultiplier','near','far','radiusX','radiusY','unk_ds3_p1_6','unk_ds3_p1_7','unk_sdt_p1_10']
       },
       [Game.EldenRing]: {
-        fields1: ['unk_sdt_f1_0','jitterAndFlicker','jitterAcceleration','unk_sdt_f1_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','shadowDarkness','unk_ds3_f1_3','unk_ds3_f1_4','fadeOutTime','unk_sdt_f1_16','unk_sdt_f1_17','unk_sdt_f1_18','volumeDensity','unk_sdt_f1_20','phaseFunction','asymmetryParam','falloffExponent','unk_er_f1_24','unk_er_f1_25','unk_er_f1_26','unk_er_f1_27'],
+        fields1: ['unk_sdt_f1_0','jitterAndFlicker','jitterAcceleration','unk_sdt_f1_3','jitterX','jitterY','jitterZ','flickerIntervalMin','flickerIntervalMax','flickerBrightness','shadows','separateSpecular','shadowDarkness','unk_ds3_f1_3','unk_ds3_f1_4','fadeOutTime','unk_sdt_f1_16','unk_sdt_f1_17','unk_sdt_f1_18','volumeDensity','unk_sdt_f1_20','phaseFunction','asymmetryParam','adaptationExponent','unk_er_f1_24','unk_er_f1_25','unk_er_f1_26','unk_er_f1_27'],
         properties1: Game.Sekiro
       },
       [Game.ArmoredCore6]: Game.EldenRing
@@ -9011,8 +9013,8 @@ const ActionDataConversion: Partial<Record<ActionType, ActionDataConversionEntry
     read(props: Props<PointLight>, game: Game) {
       props.fadeOutTime = props.fadeOutTime / 30
       if (game === Game.DarkSouls3) {
-        props.diffuseMultiplier = 100
-        props.specularMultiplier = 100
+        props.diffuseMultiplier = 10
+        props.specularMultiplier = 10
       }
       return props
     },
@@ -9022,12 +9024,12 @@ const ActionDataConversion: Partial<Record<ActionType, ActionDataConversionEntry
         const diffuseComps = separateComponents(props.diffuseColor)
         props.diffuseColor = combineComponents<ValueType.Vector4>(
           ...diffuseComps.slice(0, 3) as Vector3Components,
-          anyValueMult(diffuseComps[3], anyValueMult(1/100, props.diffuseMultiplier))
+          anyValueMult(diffuseComps[3], anyValueMult(1/10, props.diffuseMultiplier))
         ) as Vector4Value
         const specularComps = separateComponents(props.specularColor)
         props.specularColor = combineComponents<ValueType.Vector4>(
           ...specularComps.slice(0, 3) as Vector3Components,
-          anyValueMult(specularComps[3], anyValueMult(1/100, props.specularMultiplier))
+          anyValueMult(specularComps[3], anyValueMult(1/10, props.specularMultiplier))
         ) as Vector4Value
       }
       return props
@@ -9245,8 +9247,8 @@ const ActionDataConversion: Partial<Record<ActionType, ActionDataConversionEntry
         props.diffuseMultiplier = 255/100
         props.specularMultiplier = 255/100
       } else {
-        props.diffuseMultiplier = anyValueMult(1/100, props.diffuseMultiplier) as ScalarValue
-        props.specularMultiplier = anyValueMult(1/100, props.specularMultiplier) as ScalarValue
+        props.diffuseMultiplier = anyValueMult(1/10, props.diffuseMultiplier) as ScalarValue
+        props.specularMultiplier = anyValueMult(1/10, props.specularMultiplier) as ScalarValue
       }
       return props
     },
@@ -9256,8 +9258,8 @@ const ActionDataConversion: Partial<Record<ActionType, ActionDataConversionEntry
         props.diffuseColor = anyValueMult(anyValueMult(100, props.diffuseMultiplier), props.diffuseColor) as Vector4Value
         props.specularColor = anyValueMult(anyValueMult(100, props.specularMultiplier), props.specularColor) as Vector4Value
       } else {
-        props.diffuseMultiplier = anyValueMult(100, props.diffuseMultiplier) as ScalarValue
-        props.specularMultiplier = anyValueMult(100, props.specularMultiplier) as ScalarValue
+        props.diffuseMultiplier = anyValueMult(10, props.diffuseMultiplier) as ScalarValue
+        props.specularMultiplier = anyValueMult(10, props.specularMultiplier) as ScalarValue
       }
       return props
     }
@@ -14837,15 +14839,15 @@ class PeriodicEmitter extends DataAction {
    */
   interval: ScalarValue
   /**
-   * The number of particles to emit per interval. They all spawn at the same time per interval.
+   * The number of particles or nodes to emit per emission.
    * 
    * **Default**: `1`
    * 
    * **Argument**: {@link PropertyArgument.ActiveTime Active time}
    */
-  perInterval: ScalarValue
+  perEmission: ScalarValue
   /**
-   * The total number of intervals to emit particles. Once this limit is reached, the branch will be *removed from the effect*. The node and all of its descendants will be destroyed, and there are no ways to re-enable them without respawning the entire effect as they no longer exist.
+   * The total number of emissions. Once this limit is reached, the branch will be *removed from the effect*. The node and all of its descendants will be destroyed, and there are no ways to re-enable them without respawning the entire effect as they no longer exist.
    * 
    * Can be set to `-1` to disable the limit.
    * 
@@ -14853,9 +14855,9 @@ class PeriodicEmitter extends DataAction {
    * 
    * **Argument**: {@link PropertyArgument.ActiveTime Active time}
    */
-  totalIntervals: ScalarValue
+  totalEmissions: ScalarValue
   /**
-   * Maximum number of concurrent particles. Can be set to `-1` to disable the limit.
+   * Maximum number of concurrent emitted particles or nodes. Can be set to `-1` to disable the limit.
    * 
    * **Default**: `-1`
    * 
@@ -14891,13 +14893,17 @@ class EqualDistanceEmitter extends DataAction {
    */
   threshold: ScalarValue
   /**
-   * Unknown scalar.
+   * The total number of emissions. Once this limit is reached, the branch will be *removed from the effect*. The node and all of its descendants will be destroyed, and there are no ways to re-enable them without respawning the entire effect as they no longer exist.
+   * 
+   * Can be set to `-1` to disable the limit.
    * 
    * **Default**: `-1`
+   * 
+   * **Argument**: {@link PropertyArgument.ActiveTime Active time}
    */
-  unk_ds3_p1_2: ScalarValue
+  totalEmissions: ScalarValue
   /**
-   * Maximum number of concurrent particles. Can be set to -1 to disable the limit.
+   * Maximum number of concurrent emitted particles or nodes. Can be set to `-1` to disable the limit.
    * 
    * **Default**: `-1`
    * 
@@ -21145,15 +21151,12 @@ class PointLight extends DataAction {
   unk_ds3_p1_9: ScalarValue
   /**
    * Affects the falloff of the light in some way, and how the normal of surfaces affect the intensity of the light.
-   * - At 0 or negative values, this completely disabled the light.
-   * - At 1, the light behaves like you would expect.
+   * - At 0 or negative values, this completely disables the light.
+   * - At 1, the light behaves normally, like you would expect it to.
    * - At values between 0 and 1, it seemingly makes the falloff of the light over distance stronger, so the light will sooner fade to nothing.
    * - At values greater than 1, it will make the falloff weaker until near the {@link radius maximum distance}, and then it will very quickly fade to nothing. It also makes the normal of the surfaces hit by the light matter less. At very high values, anything within the radius basically becomes full bright.
    * 
    * **Default**: `1`
-   * 
-   * See also:
-   * - {@link falloffExponent}
    */
   unk_ds3_p2_0: ScalarValue
   /**
@@ -21438,16 +21441,11 @@ class PointLight extends DataAction {
    */
   asymmetryParam: number
   /**
-   * Controls the falloff exponent of the light.
+   * When not in a dark enough area, the brightness of the light will be multiplied by 0.1^α, where α is this exponent value. The light will have normal brightness when it the area it's in is sufficiently dark.
    * 
-   * Note: This is possibly something else, but the behavior is pretty similar to a falloff exponent in a few ways.
-   * 
-   * **Default**: `1`
-   * 
-   * See also:
-   * - {@link unk_ds3_p2_0}
+   * **Default**: `0`
    */
-  falloffExponent: number
+  adaptationExponent: number
   /**
    * Unknown integer.
    * 
@@ -28533,9 +28531,11 @@ class LensFlare extends DataAction {
  * ### {@link ActionType.RichModel Action 10015 - RichModel}
  * **Slot**: {@link ActionSlots.AppearanceAction Appearance}
  * 
- * Particle with a 3D model. Similar to {@link ActionType.Model Model}, but with some different options and seemingly no way to change the blend mode.
+ * Particle with a 3D model. Similar to {@link ActionType.Model Model}, but with some different options.
  * 
  * Some models only work properly with this action and not with the Model action for some unknown reason. A good example of this is the Carian greatsword model in Elden Ring (88300), which gets horribly stretched and distorted when used with the other action, but it works fine with this one.
+ * 
+ * The name is from Elden Ring's RTTI.
  */
 class RichModel extends DataAction {
   declare readonly type: ActionType.RichModel
@@ -29034,9 +29034,11 @@ class RichModel extends DataAction {
   /**
    * Unknown integer.
    * 
+   * This is somewhat similar to the `blendMode` property of other appearance actions, but the values do not match the ones in {@link BlendMode}.
+   * 
    * **Default**: `0`
    */
-  unk_er_f1_23: number
+  unkBlendMode: number
   /**
    * Unknown float.
    * 
@@ -31858,13 +31860,11 @@ class SpotLight extends DataAction {
    */
   asymmetryParam: number
   /**
-   * Controls the falloff exponent of the light.
+   * When not in a dark enough area, the brightness of the light will be multiplied by 0.1^α, where α is this exponent value. The light will have normal brightness when it the area it's in is sufficiently dark.
    * 
-   * Note: This is possibly something else, but the behavior is pretty similar to a falloff exponent in a few ways.
-   * 
-   * **Default**: `1`
+   * **Default**: `0`
    */
-  falloffExponent: number
+  adaptationExponent: number
   /**
    * Unknown integer.
    *